// src/models/ReportView.ts
// @generated
// Automatically generated. Don't change this file manually.

export default interface ReportView {
  id: number | null;

  source: number | null;

  source_id: string | null;

  date: Date | null;

  date_detail: string | null;

  word_count: number | null;

  description: string | null;

  ts: string | null;

  location: number | null;

  city: string | null;

  district: string | null;

  country: string | null;

  water: string | null;

  other: string | null;

  geoname_id: number | null;

  latitude: string | null;

  longitude: string | null;
}